requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
tqdm>=4.65.0
requests>=2.31.0
flask>=3.0.0
werkzeug>=3.0.0